// Harken-v2-release2025.08/Harken-v2-release2025.08/packages/backend/src/utils/enums/CommonEnum.ts
export enum ScreenEnums {
	INCOME_APPROACH = 'income approach',
	SALE_APPROACH = 'sale approach',
	COST_APPROACH = 'cost approach',
	AERIAL_MAP_PAGE = 'aerial map page',
	SETUP_PAGE = 'setup page',
	OVERVIEW_PAGE = 'overview page',
	IMAGES_PAGE = 'images_page',
	MAP_BOUNDARY = 'map boundary page',
	AREA_INFO = 'area info',
	EXHIBITS = 'exhibits',
	REPORT_SCREEN = 'report screen',
	LEASE_APPROACH = 'lease approach',
}

export enum ApproachesEnum {
	SALE = 'sale',
	LEASE = 'lease',
	COST = 'cost',
}
export enum RentRollEnums {
	APPRAISAL_RENT_ROLL_TYPE_ID = 'appraisal_rent_roll_type_id',
	NONE = 'None',
	BED = 'bed',
	BATH = 'bath',
	EVALUATION_RENT_ROLL_TYPE_ID = 'evaluation_rent_roll_type_id',
}
export enum EvaluationReportEnum {
	LINK = 'link',
	COMMENTS = 'comments',
	PER_SF = '/SF',
	PER_AC = '/AC',
	NOT_AVAILABLE = 'N/A',
}